Consolidate search components into combobox component
