+++ app/controllers/searches_controller.rb
@@ -1,11 +1,12 @@
 class SearchesController < ApplicationController
   def show
-    pages = if params[:query].present?
-      Page.search(params[:query]).limit(3)
+    query = params[:query] || ""
+    pages = if query.present?
+      Page.search(query).limit(3)
     else
       Page.limit(3)
     end
 
-    render Searches::ShowView.new(pages:, params:)
+    render Searches::ShowView.new(pages:, query:)
   end
 end
