+++ app/javascript/css/application.css
@@ -43,4 +43,4 @@
 @import './components/flash.css';
 @import './components/dialog.scss';
 @import './components/figures.css';
-@import './components/combolist.css';
+@import './components/combobox.css';
