+++ app/javascript/css/components/combobox.css
@@ -1,4 +1,4 @@
-.combolist {
+.combobox {
   & ul {
     list-style-type: none;
     padding: 0;
