+++ app/views/searches/show_view.rb
@@ -3,31 +3,18 @@ class Searches::ShowView < ApplicationView
   include Phlex::Rails::Helpers::Routes
   include Phlex::Rails::Helpers::TurboFrameTag
 
-  attr_reader :pages, :params
-  def initialize(pages:, params:)
+  attr_reader :pages, :query
+
+  def initialize(pages:, query:)
     @pages = pages
-    @params = params
+    @query = query
   end
 
   def view_template
     div(
       class: "section-content container py-gap mb-3xl"
     ) do
-      div(
-        class: "combolist",
-        data: {
-          controller: "combolist",
-          action: "
-            keydown->combolist#navigate
-          "
-        }
-      ) do
-        render Searches::Form.new(query: params[:query], aria: {expanded: pages.any?, controls: "search-results"})
-
-        turbo_frame_tag :search do
-          render Searches::Results.new(pages: pages, params: params, listbox: {id: "search-results"})
-        end
-      end
+      render Searches::Combobox.new(pages:, query:)
     end
   end
 end
